Chore : Simplified Verification logic
