--- client/src/components/verificationStatus/VerificationStatus.css
@@ -1,6 +1,0 @@
-.reset-success-card-wrapper {
-	margin: 100px 14px;
-	padding: 30px;
-	display: grid;
-	place-items: center;
-}
