+++ client/src/pages/verification/Verification.css
@@ -1,5 +1,6 @@
-.main {
-	display: flex;
-	justify-content: center;
-	margin: 100px 14px 80px;
+.reset-success-card-wrapper {
+	margin: 100px 14px;
+	padding: 30px;
+	display: grid;
+	place-items: center;
 }
