+++ client/src/pages/verification/Verification.test.js
@@ -1,27 +1,24 @@
 import React from 'react';
-import {render, screen, waitFor} from '@testing-library/react';
-import {BrowserRouter, MemoryRouter, Route, Routes} from 'react-router-dom';
+import {MemoryRouter, Routes, Route} from 'react-router-dom';
+import {render, waitFor, screen, act} from '@testing-library/react';
+import {useApi} from '../../hooks/useApi';
+import Home from '../welcome/Home';
 import Verification from './Verification';
-import Home from '../../pages/welcome/Home';
 
-describe('Verification component', () => {
-	test('renders VerificationStatus component when token is present', async () => {
-		const token = 'exampleToken';
-		const queryParams = new URLSearchParams();
-		queryParams.append('token', token);
-		const url = `/verify?${queryParams.toString()}`;
-		render(
-			<MemoryRouter initialEntries={[url]}>
-				<Routes>
-					<Route path='/verify' element={<Verification />} />
-				</Routes>
-			</MemoryRouter>,
-		);
-		await waitFor(() => {
-			expect(screen.getByTestId('verificationStatus')).toBeInTheDocument();
-		});
+jest.mock('../../hooks/useApi');
+
+describe('VerificationStatus Component', () => {
+	beforeEach(() => {
+		jest.clearAllMocks();
 	});
-	test('renders Home component when token is not present', async () => {
+
+	it('should nvigate to welcome page when token is not provided', async () => {
+		useApi.mockReturnValue({
+			errorMsg: null,
+			makeRequest: jest.fn(),
+			isSuccess: true,
+			loading: false,
+		});
 		const token = 'exampleToken';
 		const queryParams = new URLSearchParams();
 		queryParams.append('token', token);
@@ -36,6 +33,57 @@ describe('Verification component', () => {
 		);
 		await waitFor(() => {
 			expect(screen.getByTestId('home-container')).toBeInTheDocument();
+		});
+	});
+
+	it('renders "Email has been verified successfully" message when token is valid', async () => {
+		useApi.mockReturnValue({
+			errorMsg: null,
+			makeRequest: jest.fn(),
+			isSuccess: true,
+			loading: false,
+		});
+		const token = 'exampleToken';
+		const queryParams = new URLSearchParams();
+		queryParams.append('token', token);
+		const url = `/verify?${queryParams.toString()}`;
+		render(
+			<MemoryRouter initialEntries={[url]}>
+				<Routes>
+					<Route path='/verify' element={<Verification />} />
+					<Route path='/welcome' element={<Home />} />
+				</Routes>
+			</MemoryRouter>,
+		);
+
+		await waitFor(() => {
+			expect(
+				screen.getByText('Email verified successfully'),
+			).toBeInTheDocument();
+		});
+	});
+
+	it('renders error message message when token is invalid', async () => {
+		useApi.mockReturnValue({
+			errorMsg: 'Token Expired',
+			makeRequest: jest.fn(),
+			isSuccess: false,
+			loading: false,
+		});
+
+		const token = 'exampleToken';
+		const queryParams = new URLSearchParams();
+		queryParams.append('token', token);
+		const url = `/verify?${queryParams.toString()}`;
+		render(
+			<MemoryRouter initialEntries={[url]}>
+				<Routes>
+					<Route path='/verify' element={<Verification />} />
+				</Routes>
+			</MemoryRouter>,
+		);
+		await waitFor(() => {
+			expect(screen.getByText('Token Expired')).toBeInTheDocument();
 		});
 	});
 });
